Channels fixed & Handshake is more stable
Fixed the stability for the HandShake, it's also more strict now when to
receive/send a message
Minor changes to Diffie-Hellman
Channels are working again, this time barely any performance loss!
Some bugfix at receiving packets
And some other minor changes...

+++ SecureSocketProtocol2/Network/Messages/TCP/MsgOpenChannelResponse.cs
@@ -8,7 +8,7 @@ namespace SecureSocketProtocol2.Network.Messages.TCP
 {
     internal class MsgOpenChannelResponse : IMessage
     {
-        public ulong ConnectionId;
+        public uint ConnectionId;
         public bool success;
 
         public MsgOpenChannelResponse()
@@ -17,7 +17,7 @@ namespace SecureSocketProtocol2.Network.Messages.TCP
 
         }
 
-        public MsgOpenChannelResponse(ulong connectionId, bool success)
+        public MsgOpenChannelResponse(uint connectionId, bool success)
             : base()
         {
             this.ConnectionId = connectionId;
